Cleanup: Move impl of `#[rustc_dump_object_lifetime_defaults]`
